added index.html / now style not displaying right
