+++ index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <title>RobotIDE</title>
+  </head>
+  <body>
+    <script src="node_modules/blockly/blockly_compressed.js"></script>
+    <script src="node_modules/blockly/blocks_compressed.js"></script>
+    <script src="node_modules/blockly/msg/js/en.js"></script>
+    <script src="node_modules/blockly/javascript_compressed.js"></script>
+    <script src="lib/blocks.js"></script>
+    <script src="lib/javascript-commands.js"></script>
+    <script src="index.js"></script>
+  </body>
+</html>
