+++ src/components/blockly.js
@@ -1,5 +1,4 @@
 /* global Blockly localStorage */
-
 const widget = require('cache-element/widget')
 const html = require('choo/html')
 const sf = require('sheetify')
